Restructure project by moving pages and components into organized way
