+++ src/components/common/NavBar.jsx
@@ -8,7 +8,7 @@ import {
   MenuItems,
 } from "@headlessui/react";
 import { Bars3Icon, XMarkIcon } from "@heroicons/react/24/outline";
-import { removeItem } from "../helpers/utils/localStorage";
+import { removeItem } from "../../helpers/utils/localStorage";
 import { useNavigate } from "react-router-dom";
 
 export default function NavBar() {
