--- src/data/Data.js
@@ -1,27 +1,0 @@
-const imagesPath = [
-    {
-      id : '72f3b488-fd14-4638-9280-0cccafaa849a',
-      img : "/src/assets/images/boxImg1.png",
-      name : "Movie1",
-      publishYear : "2021"
-    },
-    {
-      id : '27541513-d21e-4a16-a00f-4435b782b361',
-      img : "/src/assets/images/boxImg2.png",
-      name : "Movie2",
-      publishYear : "2019"
-    },
-    {
-      id:'07aa56bd-25e1-4fc5-a1b8-e8363a316400',
-      img : "/src/assets/images/boxImg2.png",
-      name : "Movie3",
-      publishYear : "2022"
-    },
-    {
-      id : '41ab26cb-4e9f-4b8c-82eb-3d679e96172e',
-      img : "/src/assets/images/boxImg3.png",
-      name : "Movie4",
-      publishYear : "2023"
-    }
-  ]
-  export default imagesPath;
