--- src/pages/SignUpPage.jsx
@@ -1,288 +1,0 @@
-import React, { useState } from "react";
-import "../styles/signIn.css";
-import InputField from "../components/InputField";
-import { Link, useNavigate } from "react-router-dom";
-import { useDispatch, useSelector } from "react-redux";
-import { addUser } from "../reducers/usersAuthSlice";
-
-const SignUpPage = () => {
-  const selectors = useSelector((state) => state.usersAuth);
-  const navigate = useNavigate();
-  const dispatch = useDispatch();
-  const [error, setError] = useState({
-    username: null,
-    email: null,
-    password: null,
-    confirmPassword: null,
-  });
-
-  const [inputValue, setInputValue] = useState({
-    username: "",
-    email: "",
-    password: "",
-    confirmPassword: "",
-    permisions: {
-      view: false,
-      create: false,
-      update: false,
-      delete: false,
-    },
-  });
-
-  const handlechange = (e) => {
-    setInputValue((prevVal) => {
-      return { ...prevVal, [e.target.name]: e.target.value };
-    });
-
-    if (e.target.name == "username" && e.target.value.length == 0) {
-      setError({
-        username: "UserName is required.",
-        email: null,
-        password: null,
-        confirmPassword: null,
-      });
-      return;
-    }
-
-    if (e.target.name == "username" && e.target.value.length > 0) {
-      setError({
-        username: null,
-        email: null,
-        password: null,
-        confirmPassword: null,
-      });
-      return;
-    }
-    if (e.target.name == "email" && e.target.value.length == 0) {
-      setError({
-        username: null,
-        email: "Email is required.",
-        password: null,
-        confirmPassword: null,
-      });
-      return;
-    }
-
-    if (
-      e.target.name == "email" &&
-      !e.target.value
-        .toLowerCase()
-        .match(
-          /^(([^<>()[\]\\.,;:\s@"]+(\.[^<>()[\]\\.,;:\s@"]+)*)|.(".+"))@((\[[0-9]{1,3}\.[0-9]{1,3}\.[0-9]{1,3}\.[0-9]{1,3}\])|(([a-zA-Z\-0-9]+\.)+[a-zA-Z]{2,}))$/
-        )
-    ) {
-      setError({
-        username: null,
-        email: "Email is not valid.",
-        password: null,
-        confirmPassword: null,
-      });
-      return;
-    }
-
-    if (
-      e.target.name == "email" &&
-      e.target.value
-        .toLowerCase()
-        .match(
-          /^(([^<>()[\]\\.,;:\s@"]+(\.[^<>()[\]\\.,;:\s@"]+)*)|.(".+"))@((\[[0-9]{1,3}\.[0-9]{1,3}\.[0-9]{1,3}\.[0-9]{1,3}\])|(([a-zA-Z\-0-9]+\.)+[a-zA-Z]{2,}))$/
-        )
-    ) {
-      setError({
-        username: null,
-        email: null,
-        password: null,
-        confirmPassword: null,
-      });
-      return;
-    }
-
-    if (e.target.name == "password" && e.target.value.length == 0) {
-      setError({
-        username: null,
-        email: null,
-        password: "Password is required.",
-        confirmPassword: null,
-      });
-      return;
-    }
-    if (e.target.name == "password" && e.target.value.length > 5) {
-      setError({
-        username: null,
-        email: null,
-        password: null,
-        confirmPassword: null,
-      });
-      return;
-    }
-
-    if (
-      e.target.name == "password" &&
-      e.target.value.length > 0 &&
-      e.target.value.length < 5
-    ) {
-      setError({
-        username: null,
-        email: null,
-        password: "Password length should less than 5.",
-        confirmPassword: null,
-      });
-      return;
-    }
-
-    if (
-      e.target.name == "confirmPassword" &&
-      e.target.value != inputValue.password
-    ) {
-      setError({
-        username: null,
-        email: null,
-        password: null,
-        confirmPassword: "Confirm Password does not match with Password.",
-      });
-      return;
-    }
-    if (
-      e.target.name == "confirmPassword" &&
-      e.target.value == inputValue.password
-    ) {
-      setError({
-        username: null,
-        email: null,
-        password: null,
-        confirmPassword: null,
-      });
-      return;
-    }
-  };
-
-  const handleSignUp = (e) => {
-    e.preventDefault();
-    const valid = Validate(inputValue, setError);
-    if (valid) {
-      const findUser = selectors.find((item) => item.email == inputValue.email);
-      if (findUser) {
-        alert("Already user exist with this email");
-        return;
-      }
-      dispatch(addUser(inputValue));
-      navigate("/signin");
-      return;
-    }
-  };
-
-  return (
-    <div className="bg-[url('/src/assets/images/bg-img.jpg')] bg-bottom bg-cover bg-no-repeat h-screen flex justify-center  text-white pt-11">
-      <div className="signInBox flex flex-col justify-between gap-6">
-        <h1 className="text-5xl font-semibold text-center">Sign Up</h1>
-        <InputField
-          name="username"
-          placeholder="User Name"
-          value={inputValue.username}
-          onchange={handlechange}
-          err={error.username}
-        />
-        <InputField
-          name="email"
-          placeholder="Email"
-          value={inputValue.email}
-          onchange={handlechange}
-          err={error.email}
-        />
-        <InputField
-          name="password"
-          placeholder="Password"
-          value={inputValue.password}
-          onchange={handlechange}
-          err={error.password}
-        />
-        <InputField
-          name="confirmPassword"
-          placeholder="Confirm Password"
-          value={inputValue.confirmPassword}
-          onchange={handlechange}
-          err={error.confirmPassword}
-        />
-        <div className="flex justify-center gap-1">
-          <input type="checkbox" className="checkBox unchecked:bg-gray-200" />
-          <p className="text-center checkBoxText leading-6">Remember me</p>
-        </div>
-        <button
-          type="button"
-          className="logInButton rounded-lg font-semibold w-full p-3"
-          onClick={(e) => handleSignUp(e)}
-        >
-          Sign Up
-        </button>
-        <p className="text-xs font-semibold text-center">
-          Already have an account?
-          <Link
-            to="/signIn"
-            className="ml-1
-            decoration-blue-500 text-blue-600"
-          >
-            SignIn
-          </Link>
-        </p>
-      </div>
-    </div>
-  );
-};
-
-export default SignUpPage;
-
-function Validate(inputValue, setError) {
-  let isValid = true;
-
-  if (inputValue.username == "") {
-    setError((prev) => ({
-      ...prev,
-      username: "UserName is required.",
-    }));
-    isValid = false;
-  }
-
-  if (inputValue.email == "") {
-    setError((prev) => ({
-      ...prev,
-      email: "Email is required.",
-    }));
-    isValid = false;
-  } else if (
-    !inputValue.email
-      .toLowerCase()
-      .match(
-        /^(([^<>()[\]\\.,;:\s@"]+(\.[^<>()[\]\\.,;:\s@"]+)*)|.(".+"))@((\[[0-9]{1,3}\.[0-9]{1,3}\.[0-9]{1,3}\.[0-9]{1,3}\])|(([a-zA-Z\-0-9]+\.)+[a-zA-Z]{2,}))$/
-      )
-  ) {
-    setError((prev) => ({
-      ...prev,
-      email: "Email is not valid.",
-    }));
-    isValid = false;
-  }
-
-  if (inputValue.password == "") {
-    setError((prev) => ({
-      ...prev,
-      password: "Password is required.",
-    }));
-    isValid = false;
-  }
-
-  if (inputValue.confirmPassword == "") {
-    setError((prev) => ({
-      ...prev,
-      confirmPassword: "Confirm Password is required.",
-    }));
-    isValid = false;
-  } else if (inputValue.password != inputValue.confirmPassword) {
-    setError((prev) => ({
-      ...prev,
-      confirmPassword: "Confirm Password does not match with Password",
-    }));
-    isValid = false;
-  }
-
-  return isValid;
-}
