+++ src/pages/movies/ViewMovie.jsx
@@ -1,6 +1,6 @@
 import React, { useEffect } from "react";
-import ViewCard from "../components/ViewCard";
-import { getItem } from "../helpers/utils/localStorage";
+import ViewCard from "../../components/movies/ViewCard";
+import { getItem } from "../../helpers/utils/localStorage";
 
 const ViewMovie = () => {
   useEffect(() => {
